python-data-engineer/week4/exercises/olap_proc.py
"""
OLAP Processing
- Cube operations
- Drill-down/Roll-up
"""

import psycopg2
from psycopg2.extras import RealDictCursor
import pandas as pd
from typing import Dict, List, Any
import logging
from datetime import datetime
from dataclasses import dataclass

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

@dataclass
class Dimension:
    """OLAP dimension definition."""
    name: str
    hierarchies: Dict[str, List[str]]
    attributes: List[str]

@dataclass
class Measure:
    """OLAP measure definition."""
    name: str
    function: str
    expression: str

class OLAPProcessor:
    """OLAP processing handler."""
    
    def __init__(
        self,
        dbname: str,
        user: str,
        password: str,
        host: str = 'localhost',
        port: int = 5432
    ):
        """Initialize connection."""
        self.conn_params = {
            'dbname': dbname,
            'user': user,
            'password': password,
            'host': host,
            'port': port
        }
        self.conn = None
        self.cur = None
    
    def connect(self):
        """Establish database connection."""
        try:
            self.conn = psycopg2.connect(**self.conn_params)
            self.cur = self.conn.cursor(
                cursor_factory=RealDictCursor
            )
            logger.info("Database connection established")
        except Exception as e:
            logger.error(f"Connection error: {e}")
            raise
    
    def create_cube(
        self,
        name: str,
        fact_table: str,
        dimensions: List[Dimension],
        measures: List[Measure]
    ):
        """Create OLAP cube."""
        try:
            # Create cube metadata
            self.cur.execute("""
                CREATE TABLE IF NOT EXISTS olap_cubes (
                    cube_name VARCHAR(100) PRIMARY KEY,
                    fact_table VARCHAR(100),
                    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
                )
            """)
            
            self.cur.execute("""
                INSERT INTO olap_cubes (cube_name, fact_table)
                VALUES (%s, %s)
                ON CONFLICT (cube_name) DO NOTHING
            """, (name, fact_table))
            
            # Create dimension metadata
            self.cur.execute("""
                CREATE TABLE IF NOT EXISTS olap_dimensions (
                    cube_name VARCHAR(100),
                    dimension_name VARCHAR(100),
                    hierarchy_name VARCHAR(100),
                    level_name VARCHAR(100),
                    level_order INTEGER,
                    PRIMARY KEY (
                        cube_name,
                        dimension_name,
                        hierarchy_name,
                        level_name
                    )
                )
            """)
            
            for dim in dimensions:
                for hier_name, levels in dim.hierarchies.items():
                    for i, level in enumerate(levels):
                        self.cur.execute("""
                            INSERT INTO olap_dimensions
                            VALUES (%s, %s, %s, %s, %s)
                            ON CONFLICT DO NOTHING
                        """, (
                            name,
                            dim.name,
                            hier_name,
                            level,
                            i
                        ))
            
            # Create measure metadata
            self.cur.execute("""
                CREATE TABLE IF NOT EXISTS olap_measures (
                    cube_name VARCHAR(100),
                    measure_name VARCHAR(100),
                    function VARCHAR(50),
                    expression TEXT,
                    PRIMARY KEY (cube_name, measure_name)
                )
            """)
            
            for measure in measures:
                self.cur.execute("""
                    INSERT INTO olap_measures
                    VALUES (%s, %s, %s, %s)
                    ON CONFLICT DO NOTHING
                """, (
                    name,
                    measure.name,
                    measure.function,
                    measure.expression
                ))
            
            self.conn.commit()
            logger.info(f"Created OLAP cube: {name}")
            
        except Exception as e:
            self.conn.rollback()
            logger.error(f"Cube creation error: {e}")
            raise
    
    def execute_query(
        self,
        cube: str,
        dimensions: List[str],
        measures: List[str],
        filters: Dict = None,
        grouping_sets: List[List[str]] = None
    ):
        """Execute OLAP query."""
        try:
            # Get cube metadata
            self.cur.execute("""
                SELECT fact_table
                FROM olap_cubes
                WHERE cube_name = %s
            """, (cube,))
            
            fact_table = self.cur.fetchone()['fact_table']
            
            # Build query
            select_items = []
            group_items = []
            
            # Add dimensions
            for dim in dimensions:
                select_items.append(dim)
                group_items.append(dim)
            
            # Add measures
            for measure in measures:
                self.cur.execute("""
                    SELECT function, expression
                    FROM olap_measures
                    WHERE cube_name = %s
                    AND measure_name = %s
                """, (cube, measure))
                
                measure_meta = self.cur.fetchone()
                select_items.append(
                    f"{measure_meta['function']}({measure_meta['expression']}) "
                    f"as {measure}"
                )
            
            # Build WHERE clause
            where_clause = ""
            where_params = []
            
            if filters:
                conditions = []
                for col, value in filters.items():
                    conditions.append(f"{col} = %s")
                    where_params.append(value)
                
                if conditions:
                    where_clause = "WHERE " + \
                                 " AND ".join(conditions)
            
            # Build GROUP BY clause
            if grouping_sets:
                group_clause = """
                    GROUP BY GROUPING SETS (
                        {0}
                    )
                """.format(
                    ','.join([
                        '(' + ','.join(group) + ')'
                        for group in grouping_sets
                    ])
                )
            else:
                group_clause = """
                    GROUP BY {0}
                """.format(
                    ','.join(group_items)
                )
            
            # Execute query
            query = f"""
                SELECT
                    {','.join(select_items)}
                FROM {fact_table}
                {where_clause}
                {group_clause}
            """
            
            self.cur.execute(query, where_params)
            results = self.cur.fetchall()
            
            return results
            
        except Exception as e:
            logger.error(f"Query execution error: {e}")
            raise
    
    def drill_down(
        self,
        cube: str,
        dimension: str,
        hierarchy: str,
        current_level: str,
        measures: List[str],
        filters: Dict = None
    ):
        """Drill down operation."""
        try:
            # Get next level
            self.cur.execute("""
                SELECT level_name
                FROM olap_dimensions
                WHERE cube_name = %s
                AND dimension_name = %s
                AND hierarchy_name = %s
                AND level_order = (
                    SELECT level_order + 1
                    FROM olap_dimensions
                    WHERE cube_name = %s
                    AND dimension_name = %s
                    AND hierarchy_name = %s
                    AND level_name = %s
                )
            """, (
                cube,
                dimension,
                hierarchy,
                cube,
                dimension,
                hierarchy,
                current_level
            ))
            
            next_level = self.cur.fetchone()
            
            if next_level:
                # Execute query with next level
                dimensions = [next_level['level_name']]
                
                return self.execute_query(
                    cube,
                    dimensions,
                    measures,
                    filters
                )
            else:
                logger.warning(
                    "No further drill-down possible"
                )
                return None
            
        except Exception as e:
            logger.error(f"Drill-down error: {e}")
            raise
    
    def roll_up(
        self,
        cube: str,
        dimension: str,
        hierarchy: str,
        current_level: str,
        measures: List[str],
        filters: Dict = None
    ):
        """Roll up operation."""
        try:
            # Get previous level
            self.cur.execute("""
                SELECT level_name
                FROM olap_dimensions
                WHERE cube_name = %s
                AND dimension_name = %s
                AND hierarchy_name = %s
                AND level_order = (
                    SELECT level_order - 1
                    FROM olap_dimensions
                    WHERE cube_name = %s
                    AND dimension_name = %s
                    AND hierarchy_name = %s
                    AND level_name = %s
                )
            """, (
                cube,
                dimension,
                hierarchy,
                cube,
                dimension,
                hierarchy,
                current_level
            ))
            
            prev_level = self.cur.fetchone()
            
            if prev_level:
                # Execute query with previous level
                dimensions = [prev_level['level_name']]
                
                return self.execute_query(
                    cube,
                    dimensions,
                    measures,
                    filters
                )
            else:
                logger.warning(
                    "No further roll-up possible"
                )
                return None
            
        except Exception as e:
            logger.error(f"Roll-up error: {e}")
            raise
    
    def slice_and_dice(
        self,
        cube: str,
        dimensions: List[str],
        measures: List[str],
        slice_filters: Dict,
        dice_filters: Dict
    ):
        """Slice and dice operations."""
        try:
            # Combine filters
            filters = {**slice_filters, **dice_filters}
            
            return self.execute_query(
                cube,
                dimensions,
                measures,
                filters
            )
            
        except Exception as e:
            logger.error(
                f"Slice and dice error: {e}"
            )
            raise
    
    def disconnect(self):
        """Close database connection."""
        if self.cur:
            self.cur.close()
        if self.conn:
            self.conn.close()
            logger.info("Database connection closed")

def main():
    """Main function."""
    olap = OLAPProcessor(
        dbname='dwh',
        user='dwh_user',
        password='dwh_pass'
    )
    
    try:
        # Connect to database
        olap.connect()
        
        # Define sales cube
        time_dim = Dimension(
            name='time',
            hierarchies={
                'calendar': [
                    'year',
                    'quarter',
                    'month',
                    'day'
                ]
            },
            attributes=['date_key', 'day_of_week']
        )
        
        product_dim = Dimension(
            name='product',
            hierarchies={
                'category': [
                    'category',
                    'subcategory',
                    'product'
                ]
            },
            attributes=['product_key', 'brand']
        )
        
        measures = [
            Measure(
                name='total_sales',
                function='SUM',
                expression='amount'
            ),
            Measure(
                name='avg_quantity',
                function='AVG',
                expression='quantity'
            )
        ]
        
        # Create cube
        olap.create_cube(
            'sales_cube',
            'fact_sales',
            [time_dim, product_dim],
            measures
        )
        
        # Example query
        results = olap.execute_query(
            'sales_cube',
            ['product.category', 'time.month'],
            ['total_sales', 'avg_quantity'],
            filters={'time.year': 2024},
            grouping_sets=[
                ['product.category', 'time.month'],
                ['product.category'],
                ['time.month'],
                []
            ]
        )
        
        for row in results:
            logger.info(row)
        
    finally:
        olap.disconnect()

if __name__ == '__main__':
    main() 